requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
trafilatura>=1.6.0
